new validations for the map need to make the map one ractangular array feeding spaces with another character than make it rectang
