+++ maps/map1.cub
@@ -1,13 +1,23 @@
-NO ./sprites/wall.xpm
-SO ./sprites/wall.xpm
-WE ./sprites/wall.xpm
-EA ./sprites/wall.xpm
-C 140,125,10
-F 125,200,10
+F 76,76,76
+C 99,99,99
 
-11111111
-10000001
-10010001
-10000101
-1N000001
-11111111
+NO textures/simonkraft/cobblestone.xpm
+SO textures/simonkraft/cobblestone_01.xpm
+WE textures/simonkraft/mossy_cobblestone1.xpm
+EA textures/simonkraft/cobblestone_02.xpm
+
+
+        1111111111111111111111111
+        1000000000110000000000001
+        1011000001110000000000001
+        1001000000000000000000001
+111111111011000001110000000000001
+100000000011000001110111111111111
+11110111111111011100000010001
+11110111111111011101010010001
+11000000110101011100000010001
+10000000000000001100000010001
+10000000000000001101010010001
+11000001110101011111011110N0110
+11110111 1110101 101111010001
+11111111 1111111 111111111111
