+++ maps/map2.cub
@@ -11,15 +11,15 @@ C 225,30,0
 1001000000000000000000001
 1000000000110000000000001
 1011000001110000000000001
-100100000000000000000000111111111
-111111111011000001110000000000001
-100000000011000001110111110011111
+1001000000000000000000001111111111
+1111111110110000011100000000000001
+1000000000110000011101111100111111
 11110111111111011100000010001
 11110111111111011101010010001
 11000000110101011100000010001
 10000000000000001100000010001
 10000000000000001101010010001
-1100000111010101111101111000111
+110000011101010111110111100011
 100000000011000000000000100001111
 101100000111000000000000100000001
 100100000000000000000000100000001
@@ -38,6 +38,6 @@ C 225,30,0
 11000000110101011100000010001
 10000000000000001100000010001
 10000000000000001101010010001
-11000001110101011111011110N0111
+11000001110101011111011110N01
 11110111 1110101 101111010001
 11111111 1111111 111111111111
